Move state to a dedicated file
